consolidate the itemToPayload transform to lay the groundwork for more consolidation
