+++ src/browser/transforms.js
@@ -238,19 +238,6 @@ function userTransform(item, options, callback) {
   callback(null, newItem);
 }
 
-function itemToPayload(item, options, callback) {
-  var payloadOptions = options.payload || {};
-  if (payloadOptions.body) {
-    delete payloadOptions.body;
-  }
-
-  var data = _.extend(true, {}, item.data, payloadOptions);
-  if (item._isUncaught) {
-    data._isUncaught = true;
-  }
-  callback(null, data);
-}
-
 module.exports = {
   handleItemWithError: handleItemWithError,
   ensureItemHasSomethingToSay: ensureItemHasSomethingToSay,
@@ -260,6 +247,5 @@ module.exports = {
   addPluginInfo: addPluginInfo,
   addBody: addBody,
   scrubPayload: scrubPayload,
-  userTransform: userTransform,
-  itemToPayload: itemToPayload
+  userTransform: userTransform
 };
